refactor(public): split landing page under the line cap (#544)
Extract the home/landing page (997 lines) into cohesive marketing
sub-components under $lib/components/landing/, behavior-preserving:

- LandingHero.svelte — hero + Italian animated-letter logic + all
  hero-scoped styles (flip/warm-keyword/revel-shine/hero-warm-glow)
- LandingUseCases.svelte — the six locale-prefixed use-case cards
  (moves the #-justified eslint-disable no-navigation-without-resolve
  directives intact)
- FeeCalculatorModal.svelte — the fee-calculator dialog owning its own
  ticketPrice state + fee math (bind:open; trigger stays in parent)

SSR/SEO untouched: SeoHead placement byte-identical, no ssr override.
Landing page 997 -> 505 lines.
