+++ src/lib/components/landing/index.ts
@@ -1,1 +1,4 @@
 export { default as LandingPageTemplate } from './LandingPageTemplate.svelte';
+export { default as LandingHero } from './LandingHero.svelte';
+export { default as LandingUseCases } from './LandingUseCases.svelte';
+export { default as FeeCalculatorModal } from './FeeCalculatorModal.svelte';
