+++ src/routes/(public)/+page.svelte
@@ -3,15 +3,8 @@
 	import type { PageData } from './$types';
 	import { SeoHead } from '$lib/seo';
 	import { authStore } from '$lib/stores/auth.svelte';
-
-	interface Props {
-		data: PageData;
-	}
-	const { data }: Props = $props();
 	import * as m from '$lib/paraglide/messages.js';
 	import { getLocale } from '$lib/paraglide/runtime.js';
-	import { onMount } from 'svelte';
-	import { browser } from '$app/environment';
 	import { page } from '$app/stores';
 	import {
 		Calendar,
@@ -25,66 +18,18 @@
 		Eye,
 		Building2,
 		MapPin,
-		Calculator,
-		X,
-		ArrowRight,
-		Heart,
-		Server,
-		Lock
+		Calculator
 	} from '@lucide/svelte';
 	import Github from '$lib/components/icons/brand/Github.svelte';
+	import { LandingHero, LandingUseCases, FeeCalculatorModal } from '$lib/components/landing';
 
-	// Animated letter for Italian welcome (client-side only)
-	const letters = ['a', 'o', 'ə'] as const;
-	let currentLetterIndex = $state(0);
-	const currentLetter = $derived(letters[currentLetterIndex]);
-	let rotation = $state(0);
-	const isItalian = $derived(browser && getLocale() === 'it');
-
-	onMount(() => {
-		if (getLocale() === 'it') {
-			const interval = setInterval(() => {
-				// Always rotate forward by 180° (never backwards)
-				rotation += 180;
-
-				// At 90° (halfway point = 300ms), swap to the next letter
-				setTimeout(() => {
-					currentLetterIndex = (currentLetterIndex + 1) % letters.length;
-				}, 300);
-			}, 2000);
-
-			return () => clearInterval(interval);
-		}
-		// No cleanup needed for non-Italian locale
-		return undefined;
-	});
+	interface Props {
+		data: PageData;
+	}
+	const { data }: Props = $props();
 
 	// Fee calculator modal state
 	let showFeeCalculator = $state(false);
-	let ticketPrice = $state(20);
-
-	// Fee calculations
-	// Stripe fees: 1.5% + €0.25 for EU cards (higher for UK cards, see Stripe pricing)
-	const STRIPE_PERCENTAGE = 0.015; // 1.5%
-	const STRIPE_FIXED = 0.25; // €0.25
-
-	// Revel platform fee: 1.5% + €0.25
-	const REVEL_PERCENTAGE = 0.015; // 1.5%
-	const REVEL_FIXED = 0.25; // €0.25
-
-	const stripeFee = $derived(ticketPrice * STRIPE_PERCENTAGE + STRIPE_FIXED);
-	const revelFee = $derived(ticketPrice * REVEL_PERCENTAGE + REVEL_FIXED);
-	const totalFees = $derived(stripeFee + revelFee);
-	const organizerReceives = $derived(ticketPrice - totalFees);
-
-	function formatCurrency(amount: number): string {
-		return new Intl.NumberFormat('en-EU', {
-			style: 'currency',
-			currency: 'EUR',
-			minimumFractionDigits: 2,
-			maximumFractionDigits: 2
-		}).format(amount);
-	}
 
 	const isAuthenticated = $derived(authStore.isAuthenticated);
 	const features = $derived($page.data.features);
@@ -96,76 +41,7 @@
 
 <SeoHead config={data.seo} />
 
-<div class="hero-warm-glow">
-	<div class="container mx-auto px-4 py-16">
-		<div class="flex flex-col items-center justify-center text-center">
-			<h1 class="text-4xl font-bold tracking-tight text-foreground sm:text-6xl">
-				{#if isItalian}
-					<!-- Italian with animated letter flip -->
-					Benvenut<span class="flip-container">
-						<span class="flip-letter" style="transform: rotateY({rotation}deg)">
-							{currentLetter}
-						</span>
-					</span>
-					su <span class="revel-shine">Revel</span>
-				{:else}
-					<!-- Other languages using i18n.
-					     Rendered as real markup (not {@html}) so it re-renders on
-					     hydration and follows the active locale. See #505. -->
-					{m['home.welcomeTo']()} <span class="revel-shine">Revel</span>
-				{/if}
-			</h1>
-			<p
-				class="mt-8 flex flex-wrap items-center justify-center gap-3 text-2xl font-semibold tracking-wide text-foreground/90 sm:gap-5 sm:text-3xl md:text-4xl"
-			>
-				<span class="warm-keyword">{m['home.keywordFree']()}</span>
-				<span class="text-foreground/30" aria-hidden="true">&middot;</span>
-				<span class="warm-keyword">{m['home.keywordOpen']()}</span>
-				<span class="text-foreground/30" aria-hidden="true">&middot;</span>
-				<span class="warm-keyword">{m['home.keywordSafe']()}</span>
-			</p>
-			<div class="mt-10 flex flex-wrap items-center justify-center gap-4">
-				{#if isAuthenticated}
-					<a
-						href={resolve('/(auth)/dashboard', {})}
-						class="rounded-md bg-primary px-6 py-3 text-base font-semibold text-primary-foreground shadow-sm hover:bg-primary/90 focus-visible:outline focus-visible:outline-2 focus-visible:outline-offset-2 focus-visible:outline-primary"
-					>
-						{m['userMenu.dashboard']()}
-					</a>
-					<a
-						href={resolve('/(public)/events', {})}
-						class="rounded-md border border-primary px-6 py-3 text-base font-semibold text-primary hover:bg-primary/10 focus-visible:outline focus-visible:outline-2 focus-visible:outline-offset-2 focus-visible:outline-primary"
-					>
-						{m['nav.browseEvents']()}
-					</a>
-				{:else}
-					<a
-						href={resolve('/(public)/events', {})}
-						class="rounded-md bg-primary px-6 py-3 text-base font-semibold text-primary-foreground shadow-sm hover:bg-primary/90 focus-visible:outline focus-visible:outline-2 focus-visible:outline-offset-2 focus-visible:outline-primary"
-					>
-						{m['nav.browseEvents']()}
-					</a>
-					<a
-						href={resolve('/(public)/register', {})}
-						class="rounded-md border border-primary px-6 py-3 text-base font-semibold text-primary hover:bg-primary/10 focus-visible:outline focus-visible:outline-2 focus-visible:outline-offset-2 focus-visible:outline-primary"
-					>
-						{m['home.getStarted']()}
-					</a>
-				{/if}
-			</div>
-			{#if !isAuthenticated}
-				<div class="mt-4">
-					<a
-						href={resolve('/(public)/login', {})}
-						class="text-sm text-muted-foreground hover:text-foreground hover:underline"
-					>
-						{m['home.alreadyHaveAccount']()}
-					</a>
-				</div>
-			{/if}
-		</div>
-	</div>
-</div>
+<LandingHero {isAuthenticated} />
 
 <div class="container mx-auto px-4">
 	<!-- Features Section -->
@@ -577,153 +453,7 @@
 	</div>
 
 	<!-- Use Cases Section -->
-	<div class="mb-16 mt-8">
-		<div class="text-center">
-			<h2 class="mb-4 text-3xl font-bold">{m['home.useCasesTitle']()}</h2>
-			<p class="mx-auto mb-8 max-w-2xl text-muted-foreground">
-				{m['home.useCasesDescription']()}
-			</p>
-		</div>
-		<div class="grid gap-4 sm:grid-cols-2 lg:grid-cols-3">
-			<!-- eslint-disable svelte/no-navigation-without-resolve -- locale-prefixed landing path; the prefix comes from getLocale() and cannot map to a single static route id -->
-			<a
-				href="{landingPagePrefix}/queer-event-management"
-				class="group rounded-lg border bg-card p-6 transition-all hover:border-primary hover:shadow-md"
-			>
-				<div class="mb-3 flex items-center gap-3">
-					<div class="rounded-full bg-pink-100 p-2 dark:bg-pink-900/30">
-						<Heart class="h-5 w-5 text-pink-600 dark:text-pink-400" aria-hidden="true" />
-					</div>
-					<h3 class="font-semibold">{m['footer.solutionQueer']()}</h3>
-				</div>
-				<p class="mb-3 text-sm text-muted-foreground">
-					{m['home.useCaseQueerDescription']()}
-				</p>
-				<span
-					class="inline-flex items-center gap-1 text-sm font-medium text-primary group-hover:underline"
-				>
-					{m['footer.learnMore']()}
-					<ArrowRight class="h-4 w-4" aria-hidden="true" />
-				</span>
-			</a>
-			<!-- eslint-enable svelte/no-navigation-without-resolve -->
-
-			<!-- eslint-disable svelte/no-navigation-without-resolve -- locale-prefixed landing path; the prefix comes from getLocale() and cannot map to a single static route id -->
-			<a
-				href="{landingPagePrefix}/kink-event-ticketing"
-				class="group rounded-lg border bg-card p-6 transition-all hover:border-primary hover:shadow-md"
-			>
-				<div class="mb-3 flex items-center gap-3">
-					<div class="rounded-full bg-purple-100 p-2 dark:bg-purple-900/30">
-						<Shield class="h-5 w-5 text-purple-600 dark:text-purple-400" aria-hidden="true" />
-					</div>
-					<h3 class="font-semibold">{m['footer.solutionKink']()}</h3>
-				</div>
-				<p class="mb-3 text-sm text-muted-foreground">
-					{m['home.useCaseKinkDescription']()}
-				</p>
-				<span
-					class="inline-flex items-center gap-1 text-sm font-medium text-primary group-hover:underline"
-				>
-					{m['footer.learnMore']()}
-					<ArrowRight class="h-4 w-4" aria-hidden="true" />
-				</span>
-			</a>
-			<!-- eslint-enable svelte/no-navigation-without-resolve -->
-
-			<!-- eslint-disable svelte/no-navigation-without-resolve -- locale-prefixed landing path; the prefix comes from getLocale() and cannot map to a single static route id -->
-			<a
-				href="{landingPagePrefix}/privacy-focused-events"
-				class="group rounded-lg border bg-card p-6 transition-all hover:border-primary hover:shadow-md"
-			>
-				<div class="mb-3 flex items-center gap-3">
-					<div class="rounded-full bg-green-100 p-2 dark:bg-green-900/30">
-						<Lock class="h-5 w-5 text-green-600 dark:text-green-400" aria-hidden="true" />
-					</div>
-					<h3 class="font-semibold">{m['footer.solutionPrivacy']()}</h3>
-				</div>
-				<p class="mb-3 text-sm text-muted-foreground">
-					{m['home.useCasePrivacyDescription']()}
-				</p>
-				<span
-					class="inline-flex items-center gap-1 text-sm font-medium text-primary group-hover:underline"
-				>
-					{m['footer.learnMore']()}
-					<ArrowRight class="h-4 w-4" aria-hidden="true" />
-				</span>
-			</a>
-			<!-- eslint-enable svelte/no-navigation-without-resolve -->
-
-			<!-- eslint-disable svelte/no-navigation-without-resolve -- locale-prefixed landing path; the prefix comes from getLocale() and cannot map to a single static route id -->
-			<a
-				href="{landingPagePrefix}/self-hosted-event-platform"
-				class="group rounded-lg border bg-card p-6 transition-all hover:border-primary hover:shadow-md"
-			>
-				<div class="mb-3 flex items-center gap-3">
-					<div class="rounded-full bg-blue-100 p-2 dark:bg-blue-900/30">
-						<Server class="h-5 w-5 text-blue-600 dark:text-blue-400" aria-hidden="true" />
-					</div>
-					<h3 class="font-semibold">{m['footer.solutionSelfHosted']()}</h3>
-				</div>
-				<p class="mb-3 text-sm text-muted-foreground">
-					{m['home.useCaseSelfHostedDescription']()}
-				</p>
-				<span
-					class="inline-flex items-center gap-1 text-sm font-medium text-primary group-hover:underline"
-				>
-					{m['footer.learnMore']()}
-					<ArrowRight class="h-4 w-4" aria-hidden="true" />
-				</span>
-			</a>
-			<!-- eslint-enable svelte/no-navigation-without-resolve -->
-
-			<!-- eslint-disable svelte/no-navigation-without-resolve -- locale-prefixed landing path; the prefix comes from getLocale() and cannot map to a single static route id -->
-			<a
-				href="{landingPagePrefix}/eventbrite-alternative"
-				class="group rounded-lg border bg-card p-6 transition-all hover:border-primary hover:shadow-md"
-			>
-				<div class="mb-3 flex items-center gap-3">
-					<div class="rounded-full bg-orange-100 p-2 dark:bg-orange-900/30">
-						<Ticket class="h-5 w-5 text-orange-600 dark:text-orange-400" aria-hidden="true" />
-					</div>
-					<h3 class="font-semibold">{m['footer.solutionEventbrite']()}</h3>
-				</div>
-				<p class="mb-3 text-sm text-muted-foreground">
-					{m['home.useCaseEventbriteDescription']()}
-				</p>
-				<span
-					class="inline-flex items-center gap-1 text-sm font-medium text-primary group-hover:underline"
-				>
-					{m['footer.learnMore']()}
-					<ArrowRight class="h-4 w-4" aria-hidden="true" />
-				</span>
-			</a>
-			<!-- eslint-enable svelte/no-navigation-without-resolve -->
-
-			<!-- eslint-disable svelte/no-navigation-without-resolve -- locale-prefixed landing path; the prefix comes from getLocale() and cannot map to a single static route id -->
-			<a
-				href="{landingPagePrefix}/community-first-event-platform"
-				class="group rounded-lg border bg-card p-6 transition-all hover:border-primary hover:shadow-md"
-			>
-				<div class="mb-3 flex items-center gap-3">
-					<div class="rounded-full bg-teal-100 p-2 dark:bg-teal-900/30">
-						<Users class="h-5 w-5 text-teal-600 dark:text-teal-400" aria-hidden="true" />
-					</div>
-					<h3 class="font-semibold">{m['footer.solutionCommunity']()}</h3>
-				</div>
-				<p class="mb-3 text-sm text-muted-foreground">
-					{m['home.useCaseCommunityDescription']()}
-				</p>
-				<span
-					class="inline-flex items-center gap-1 text-sm font-medium text-primary group-hover:underline"
-				>
-					{m['footer.learnMore']()}
-					<ArrowRight class="h-4 w-4" aria-hidden="true" />
-				</span>
-			</a>
-			<!-- eslint-enable svelte/no-navigation-without-resolve -->
-		</div>
-	</div>
+	<LandingUseCases {landingPagePrefix} />
 
 	<!-- Try the Demo Section -->
 	<div class="mb-16 rounded-lg border-2 border-secondary bg-secondary/5 p-8 text-center">
@@ -772,226 +502,4 @@
 </div>
 
 <!-- Fee Calculator Modal -->
-{#if showFeeCalculator}
-	<div
-		class="fixed inset-0 z-50 flex items-center justify-center bg-black/50 p-4"
-		onclick={(e) => e.target === e.currentTarget && (showFeeCalculator = false)}
-		onkeydown={(e) => e.key === 'Escape' && (showFeeCalculator = false)}
-		role="dialog"
-		aria-modal="true"
-		aria-labelledby="fee-calculator-title"
-		tabindex="-1"
-	>
-		<div class="w-full max-w-md rounded-lg border bg-background shadow-xl">
-			<!-- Header -->
-			<div class="flex items-center justify-between border-b px-6 py-4">
-				<h2 id="fee-calculator-title" class="text-xl font-bold">
-					{m['learnMore.feeCalculator.title']()}
-				</h2>
-				<button
-					type="button"
-					onclick={() => (showFeeCalculator = false)}
-					class="rounded-full p-1 hover:bg-accent"
-					aria-label={m['learnMore.feeCalculator.close']()}
-				>
-					<X class="h-5 w-5" />
-				</button>
-			</div>
-
-			<!-- Content -->
-			<div class="space-y-6 px-6 py-4">
-				<!-- Ticket Price Input -->
-				<div>
-					<label for="ticket-price" class="mb-2 block text-sm font-medium"
-						>{m['learnMore.feeCalculator.ticketPrice']()}</label
-					>
-					<div class="relative">
-						<span class="absolute left-3 top-1/2 -translate-y-1/2 text-muted-foreground">€</span>
-						<input
-							id="ticket-price"
-							type="number"
-							min="0"
-							step="0.01"
-							bind:value={ticketPrice}
-							class="w-full rounded-md border border-input bg-background py-2 pl-8 pr-4 text-lg focus:outline-none focus:ring-2 focus:ring-primary"
-						/>
-					</div>
-				</div>
-
-				<!-- Fee Breakdown -->
-				<div class="space-y-4">
-					<!-- Stripe Fees -->
-					<div class="rounded-lg border bg-muted/30 p-4">
-						<div class="flex items-center justify-between">
-							<span class="font-medium">{m['learnMore.feeCalculator.creditCardFees']()}</span>
-							<span class="text-lg font-bold text-orange-600 dark:text-orange-400">
-								{formatCurrency(stripeFee)}
-							</span>
-						</div>
-						<p class="mt-1 text-xs text-muted-foreground">
-							{m['learnMore.feeCalculator.creditCardFeesDescription']()}
-							<a
-								href="https://stripe.com/en-at/pricing"
-								target="_blank"
-								rel="noopener noreferrer"
-								class="text-primary hover:underline"
-							>
-								{m['learnMore.feeCalculator.viewStripePricing']()}
-							</a>
-						</p>
-					</div>
-
-					<!-- Revel Platform Fee -->
-					<div class="rounded-lg border bg-muted/30 p-4">
-						<div class="flex items-center justify-between">
-							<span class="font-medium">{m['learnMore.feeCalculator.platformFee']()}</span>
-							<span class="text-lg font-bold text-primary">
-								{formatCurrency(revelFee)}
-							</span>
-						</div>
-						<p class="mt-1 text-xs text-muted-foreground">
-							{m['learnMore.feeCalculator.platformFeeDescription']()}
-						</p>
-					</div>
-
-					<!-- Divider -->
-					<div class="border-t"></div>
-
-					<!-- Organization Receives -->
-					<div class="rounded-lg border-2 border-green-500 bg-green-50 p-4 dark:bg-green-950/30">
-						<div class="flex items-center justify-between">
-							<span class="font-semibold"
-								>{m['learnMore.feeCalculator.organizationReceives']()}</span
-							>
-							<span class="text-2xl font-bold text-green-600 dark:text-green-400">
-								{formatCurrency(organizerReceives)}
-							</span>
-						</div>
-						<p class="mt-1 text-xs text-muted-foreground">
-							{m['learnMore.feeCalculator.perTicketSoldAt']({ price: formatCurrency(ticketPrice) })}
-						</p>
-					</div>
-				</div>
-
-				<!-- Summary -->
-				<p class="text-center text-xs text-muted-foreground">
-					{m['learnMore.feeCalculator.totalFees']({
-						fees: formatCurrency(totalFees),
-						percentage: ((totalFees / ticketPrice) * 100).toFixed(1)
-					})}
-				</p>
-				<p class="text-center text-xs text-muted-foreground/70">
-					{m['learnMore.feesExcludeVat']()}
-				</p>
-			</div>
-
-			<!-- Footer -->
-			<div class="border-t px-6 py-4">
-				<p class="text-center text-sm text-muted-foreground">
-					{m['learnMore.feeCalculator.questionsAboutFees']()}
-					<a
-						href="mailto:contact@letsrevel.io?subject=Revel%20Fee%20Question"
-						class="text-primary hover:underline"
-					>
-						{m['learnMore.feeCalculator.contactUs']()}
-					</a>
-				</p>
-			</div>
-		</div>
-	</div>
-{/if}
-
-<style>
-	.flip-container {
-		perspective: 1000px;
-		display: inline-block;
-	}
-
-	.flip-letter {
-		display: inline-block;
-		transition: transform 0.6s ease-in-out;
-		transform-style: preserve-3d;
-	}
-
-	.hero-warm-glow {
-		position: relative;
-		overflow: hidden;
-	}
-
-	.hero-warm-glow::before {
-		content: '';
-		position: absolute;
-		top: -40%;
-		left: 50%;
-		transform: translateX(-50%);
-		width: 80%;
-		height: 100%;
-		background: radial-gradient(
-			ellipse at center,
-			hsl(30 80% 55% / 0.06) 0%,
-			hsl(350 70% 55% / 0.03) 40%,
-			transparent 70%
-		);
-		pointer-events: none;
-		z-index: 0;
-	}
-
-	.hero-warm-glow > :global(*) {
-		position: relative;
-		z-index: 1;
-	}
-
-	.warm-keyword {
-		color: hsl(30 80% 65%);
-	}
-
-	@supports ((-webkit-background-clip: text) or (background-clip: text)) {
-		.warm-keyword {
-			background: linear-gradient(135deg, hsl(30 80% 70%) 0%, hsl(350 65% 65%) 100%);
-			-webkit-background-clip: text;
-			-webkit-text-fill-color: transparent;
-			background-clip: text;
-		}
-	}
-
-	:global(.revel-shine) {
-		display: inline-block;
-		color: hsl(var(--primary));
-	}
-
-	@supports ((-webkit-background-clip: text) or (background-clip: text)) {
-		:global(.revel-shine) {
-			background: linear-gradient(
-				105deg,
-				hsl(var(--primary)) 0%,
-				hsl(var(--primary)) 40%,
-				hsl(30 80% 70%) 48%,
-				hsl(350 65% 70%) 52%,
-				hsl(var(--primary)) 60%,
-				hsl(var(--primary)) 100%
-			);
-			background-size: 400% 100%;
-			background-position: 100% 0;
-			-webkit-background-clip: text;
-			-webkit-text-fill-color: transparent;
-			background-clip: text;
-			animation: shine-glare 7.5s ease-in-out infinite;
-		}
-	}
-
-	@media (prefers-reduced-motion: reduce) {
-		:global(.revel-shine) {
-			animation: none;
-		}
-	}
-
-	@keyframes shine-glare {
-		0%,
-		67% {
-			background-position: 100% 0;
-		}
-		100% {
-			background-position: 0% 0;
-		}
-	}
-</style>
+<FeeCalculatorModal bind:open={showFeeCalculator} />
